requirements.txt
# Core web framework
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6

# Supabase integration
supabase>=2.17.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Amazon Advertising API
requests==2.31.0
requests-oauthlib==1.3.1
boto3==1.34.25

# Authentication and security
pyjwt>=2.10.1
cryptography>=41.0.7
passlib==1.7.4
bcrypt==4.1.2

# HTTP client for API calls
httpx>=0.26.0
tenacity==8.2.3

# Environment and config
python-dotenv==1.0.0
pyyaml==6.0.1

# Date/time handling
python-dateutil==2.8.2
croniter==2.0.1

# Rate limiting
ratelimit==2.2.1
slowapi==0.1.9

# Testing (optional)
pytest==7.4.4
pytest-asyncio==0.23.3

# Development tools (optional)
black==23.12.1
flake8==7.0.0

# Monitoring (optional)
prometheus-client==0.19.0
sentry-sdk==1.39.2